Move ExchangeDeclareStrategy to root namespace
It is quite high level abstraction, not related to other low producer things
